Adds one line to the "Adding to this list" section
By browsing through the PRs that we're getting on a daily basis, it seems like it would be easier to request from the people to submit a link directly in the body of the commit message. That way, we can just click on it instead of switching to the `Files` tab, copying it and then pasting it in a new browser tab.

It's just something that came to my mind like right now, so I'm gonna wait for the feedback first.